Bug 1903974 On frame removal, subtract only new gaps from mSanitizedBufferedRanges r=media-playback-reviewers,alwu
The replaced code that recalculated mSanitizedBufferedRanges with a different
fuzz would introduce additional spurious gaps in the sanitized buffered ranges
when the durations of the frames removed were small.  maxSampleDuration was
intended to apply only as fuzz to removed frames [1].  InsertFrames() uses the
longest frame duration since the key frame [2] to fill in small gaps in
reported buffered ranges, which was being undone by the recalculation with a
different fuzz.  When the removed duration was zero, no fuzz at all was used
in the recalculation.

[1]
https://hg.mozilla.org/mozilla-central/rev/849e4c4784ca925c55ffeb283fca1a5ad13f8079#l1.58
[2]
https://searchfox.org/mozilla-central/rev/7f7ea528f740dfbfd871c28362070982ef5f01e9/dom/media/mediasource/TrackBuffersManager.cpp#2153,2155-2157,2352-2353

Differential Revision: https://phabricator.services.mozilla.com/D214517

UltraBlame original commit: 53b744845abd6c4604e42931ca0287eb36dbc9e5