Bug 1894512 - Add an AndroidView composable that supports properly positioning a menu r=android-reviewers,007
A fully Compose implementation that would use a DropdownMenu wouldn’t allow positioning
the menu to the bottom of the screen because DropdownMenu use hardcoded margins and paddings.
Following community requests this shortcoming is to be fixed in
https://github.com/JetBrains/compose-multiplatform/issues/1878 after which we can migrate
to a fully Compose implementation.

Differential Revision: https://phabricator.services.mozilla.com/D214360

UltraBlame original commit: 64d8e75c61fa835626e6b9e2ba8975614efdb3c4