Bug 1898153 - wasm: Refactor ModuleGenerator for supporting multiple code blocks. r=jseward
This commit reworks the logic in ModuleGenerator so that it's ready
to generate multiple code blocks. The existing code path for a
single 'complete tier' is made into a new method.

  * ModuleGenerator::masm_ is moved to a separate heap allocated struct. This allows
    the masm_ to be created new for each code block.
  * ModuleGenerator::startBlock creates a new CodeBlock that all compile functions
    implicitly operate on. (e.g. compileFuncDef, linkCompiledCode). This asserst
    that we are not already inside of an actively generated CodeBlock.
  * ModuleGenerator::finishBlock finishes the CodeBlock and returns it. All our
    debug assertions about final generated code are moved here.
  * ModuleGenerator::startCompleteTier starts a new block and initializes it with
    the metadata we expect to have.
  * ModuleGenerator::finishCompleteTier ends the compilation of functions, and
    generates stubs, then finishes the blocks.

A future commit will use this logic to generate a 'shared stubs' block which
contain stubs and import functions that are tier independent. After that, we
can extend this further with startPartialTier and finishPartialTier methods.

Differential Revision: https://phabricator.services.mozilla.com/D213150

UltraBlame original commit: e36dc8935bc06f90fa13e55aba1ec7925a689cf5