Bug 1905011 - Add testing for InitializeStorage scheduled just before ShutdownStorage is finished; r=dom-storage-reviewers,jari
This is an edge case and was the primary reason for making DirectoryLock::Drop
async. Async Drop is needed for replacing counting of pending clear/shutdown storage operations with evaluation of existing directory locks because of this
scenario:
- ShutdownStorageOp::DoDirectoryWork is done
- NormalOriginOperationBase::UnblockOpen calls SendResults
- ResolvableNormalOriginOp::SendResults resolves the promise (a new runnable is
  dispatched to the current thread)
- NormalOriginOperationBase::UnblockOpen calls CloseDirectory
- ShutdownStorageOp::CloseDirectory drops the directory lock immediately and
  synchronously
- The current runnable is done
- A pending runnable is being processed which calls InitializeStorage
- InitializeStorage still sees that mStorageInitializes is true and that there
  are no existing directory locks, so it immediately resolves its promise
- The runnable dispatched from ResolvableNormalOriginOp::SendResults is
  processed and the handler for it sets mStorageInitialized to false in
  QuotaManager::ShutdownStorage
- Finally, mStorageInitialized is false, but it should be true

So the problem in the scenario is that the directory lock is dropped
synchronously. However, all directory locks are now dropped explicitly and
asynchronously so InitializeStorage would still see the directory lock for
ShutdownStorage and wouldn't immediately resolve its promise.

This new test simulates the described scenario.

Differential Revision: https://phabricator.services.mozilla.com/D197322

UltraBlame original commit: ebd5ed5b9adc2055a902bee884aeac3f41a1282a