Bug 1906026 - Make dependentlibs.py fail when it doesn't find any dependencies. r=firefox-build-system-reviewers,nalexander
Until such a time where libxul doesn't have dependencies at all, this is
not supposed to happen. When it does, it's because the output from the
tool used to produce the list changed, or the tool is broken is some
other way.

Differential Revision: https://phabricator.services.mozilla.com/D215582

UltraBlame original commit: 5cccb9c5d12e12c0d9d78242fa92fda3e5681a67