Bug 1905604 - Fix edge case with registered color computation. r=firefox-style-system-reviewers,devtools-reviewers,nchevobbe,zrhoffman
The issue is that we track the 1em as a custom reference and thus fail
to compute the color. But in this case 1em is not a valid value and we
should fall back to the initial value.

Differential Revision: https://phabricator.services.mozilla.com/D216030

UltraBlame original commit: d33a0841dc39b2de3ee177de880bd643b32f6c24