Bug 1897328 Replace integral controller term with clock drift estimate r=pehrsons
The drift estimate is not affected by feedback from the PD controller and so
does not have the same stability risks as the integral term, such as integral
wind up while the change in response is being capped.  The response while not
capped and outside hysteresis is now purely an exponential decay, no longer
with an oscillatory component.  When the capping is applied, the lag in
adjustment still causes some oscillations from overshoot, but these now
attenuate much faster.

If the drift estimate applied to the corrected rate is accurate enough, then
the desired steady state contribution from the remaining terms approaches zero
as the error approaches zero, so the integral term is no longer required.

The rate at which the drift estimate adjusts to changes in input depends on
packet rate.  For typical low latency packet sizes, the drift estimate should
adjust faster than the integral term.  This makes the controller more
resistant to underrun in TestAudioDriftCorrection.DriftStepResponseUnderrun,
which is adjusted according.

The clock drift estimate does not anticipate ongoing desired buffer level
changes like the intergral term did, and so
TestAudioDriftCorrection.DynamicInputBufferSizeChanges does not reach the
hysteresis band by the end of the test.

TestDriftController.VerySmallBufferedFrames now runs slightly longer because
the rate of change in correction is not quite as steep as the capped rate.

Differential Revision: https://phabricator.services.mozilla.com/D215487

UltraBlame original commit: 545e985e24285cc382d9ebf51bf7edb42e5be2bd