Bug 1907151 - wasm: Only generate shared stubs when compiling tier 1. r=jseward
ModuleGenerator used to compile the shared stubs every time it was
initialized to compile a complete tier. This is wasteful, because it
would be thrown away when compiling tier2. However, we need the shared
stubs link data for serializing tier2, making fixing this tricky.

This commit reworks serialization so that the LinkData that we
serialize comes from wasm::Code. This lets tier1 put the shared
stubs link data into wasm::Code, then have the tier2 compile
pull it out when it's time to serialize.

With that in place, we can now skip compiling shared stubs when
we're not in tier1.

Differential Revision: https://phabricator.services.mozilla.com/D216220

UltraBlame original commit: 67a952c44d8995a8f2f548eb5f6a1687067f8bc3