Bug 1831649 - Explicitly set android.windowSoftInputMode to GeckoView test runner. r=geckoview-reviewers,m_kato
And add a caveat about windowSoftInputMode in a document.

We implement interactive-widget [1] in a way that the content is laid out as if
the window doesn't get resized even if the application window gets resized by
the software keyboard. To make this way work, using `adjustResize` is mandatory.

An alternative approach to implement it is to change windowSoftInputMode value
depending on each content's interactive-widget value in each Android Activity,
i.e., set to `adjustResize` on `interactive-widget=resizes-content` mode, set to
`adjustUnspecified` on `interactive-widget=resizes-visual` or on
`interactive-widget=overlays-mode`. It will force applications to do more work.

[1] https://drafts.csswg.org/css-viewport/#interactive-widget-section

Differential Revision: https://phabricator.services.mozilla.com/D204165

UltraBlame original commit: 22538f7bdc889ccb3ddb6e0e943ebfa818c3e9e6