Bug 1898238 - Keep SharedSurface_EGLImages alive but do not recycle them. r=sotaro,jgilbert
We must ensure that a SharedSurface_EGLImage outlives its
corresponding RenderEGLImageTextureHost, so that
RenderEGLImageTextureHost does not attempt to wait on an EGLSync that
has been destroyed. This can be achieved by passing the surface to
RemoteTextureOwnerClient::PushTexture(), ensuring the RemoteTextureMap
keeps it alive.

However, we must be careful not to recycle the SharedSurface_EGLImage,
as that is currently not supported. To achieve this, we make
RemoteTextureMap::RecycleTexture() only recycle SharedSurfaces whose
mDesc.canRecycle value is true.

Differential Revision: https://phabricator.services.mozilla.com/D211292

UltraBlame original commit: 549abd685d117a8b359bbbaa6b3321ccb35622a6