Bug 1892021, part 3 - Credential chosen callback update to match new semantic - r=anti-tracking-reviewers,pbz
Down the stack I need to change the argument to `notify` in the nsICredentialChosenCallback so it
can run in Javascript for testing. I make the change here that allows consumers of this class to
still get the `IPCIdentityCredential` in the resolve callback of the argument promise.

Differential Revision: https://phabricator.services.mozilla.com/D215007

UltraBlame original commit: 5d0683eaf762101c0bb0f276d23fdc6fb5496c00