Bug 1857021 - Part 3: Make the FxaWebChannelFeature dismiss the activity on logout r=android-reviewers,amejiamarmol
To connect the feature to the UI, we wrap it in an integration class
so that its isolated as we start to handle more commands.

We also early-return when we know that the shared UI is opened from
within the app instead of from a third party app.

Differential Revision: https://phabricator.services.mozilla.com/D217159

UltraBlame original commit: fa0ced592e7df8a0e1adcd470c0227c140902c6a