Bug 1857021 - Part 4: Add unit tests to FxaWebChannelFeature r=android-reviewers,amejiamarmol
When adding some of these tests, we add a `MainCoroutineRule` which
now makes anything that was wrapped in the a `Dispatchers.Main`
coroutine. Since we weren't doing this before, a bunch of tests
started to fail that were not exercising the real codepaths.

Another example where we need to stop using mocks everywhere unless
it is necessary.

Differential Revision: https://phabricator.services.mozilla.com/D217160

UltraBlame original commit: 82c5d07c2f165711f69d37fba421914a2d1ce1a8