Bug 1907945 - Support labeled_timing_distribution metrics in FOG r=TravisLong
This one is a little more complex because we need to support GIFFT for pure
Rust consumers (bug 1751448).

To do that we need to split the GIFFT impl from the core impl of FOG's Rust
timing_distribution implementation. That makes it so that the labeled_*-
specific submetric impl can dispatch to the core impl and call to GIFFT
independently.

It also makes the GIFFT code in timing_distribution.rs a little nicer
(to my eyes), which is a nice plus.

Differential Revision: https://phabricator.services.mozilla.com/D217611

UltraBlame original commit: cfa2a54bceaefd93b284e5b44aeaba41a47a5608