Bug 1862911 - Fix restartDownloadFromAppNotificationAfterConnectionIsInterruptedTest UI test r=ohorvath
The UI test was disabled because sometimes the download failed prompt was displayed later than usual. (did a status check and noticed that sometimes it's being displayed right away, in other cases it was displayed with a delay)

To overcome this problem I've added a retry loop to extend the waiting time for it to be displayed when needed.

The UI test successfully passed 150x on Firebase.

Differential Revision: https://phabricator.services.mozilla.com/D217555

UltraBlame original commit: 8a2f901ce5d48d4003993c336c703e6807082bdc