Bug 1910413 - Introduce a minimum time clamp for the "age" of a report in Reporting API r=dom-core,jjaschke
After https://phabricator.services.mozilla.com/D202879 is landed,
sending a report in Reporting API no longer needs to wait for
a timeout, it can be sent right away when the event queue is Idle.
This can potentially cause the "age" of a report to be 0.

So here we introduce a clamp for it to make it non-zero.

Differential Revision: https://phabricator.services.mozilla.com/D218021

UltraBlame original commit: 4559aff9fa7e2b418d0f5a637f634920af33b67e