Bug 1901586 - Uplift web platform tests for Lightweight FedCM from mozilla folder - r=pbz
These tests rely on the test_driver.select_fedcm_account endpoint that we don't support.
But it sort-of automagically works because we have the wpt-pick-me autoselect behavior pref'd in.
We are going to get rid of that in Bug 1910688.

Differential Revision: https://phabricator.services.mozilla.com/D218098

UltraBlame original commit: 8cd6dcd484dfc6bb00832b354ab50f5ec2038fe2